feat(db+a): make interface & relational configs multifile
