+++ common/relational/__init__.py
@@ -1,37 +1,9 @@
 from __future__ import annotations
 
-from sqlalchemy import (
-    create_engine,
-    select,
-    MetaData,
-    Column,
-    Integer,
-    String,
-)
-
+from .config_db import configure_sqlalchemy
 from .sqlalchemy_ext import Sessionmaker, Session, create_base
-from ..interface import (
-    Database,
-    from_orm,
-    User,
-)
-
-
-def configure_sqlalchemy(db_url: str):
-    convention = {
-        "ix": "ix_%(column_0_label)s",
-        "uq": "uq_%(table_name)s_%(column_0_name)s",
-        "ck": "ck_%(table_name)s_%(constraint_name)s",
-        "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
-        "pk": "pk_%(table_name)s"
-    }
-
-    engine = create_engine(db_url, pool_recycle=280)  # echo=True
-    db_meta = MetaData(bind=engine, naming_convention=convention)
-    Base = create_base(db_meta)
-    sessionmaker = Sessionmaker(bind=engine, class_=Session)
-
-    return db_meta, Base, sessionmaker
+from .users_rdb import build_user_database
+from ..interface import Database
 
 
 def build_sqlalchemy_database(db_url: str, *config) -> Database:
@@ -39,49 +11,10 @@ def build_sqlalchemy_database(db_url: str, *config) -> Database:
         config = configure_sqlalchemy(db_url)
     db_meta, Base, sessionmaker = config
 
-    class UserORM(Base):
-        __tablename__ = "users"
+    UserDatabase = build_user_database(Base, sessionmaker)
 
-        id = Column(Integer, primary_key=True)
-        username = Column(String(100), nullable=False, index=True)
-        password = Column(String(100), nullable=False)
-
-    class BlockedToken(Base):
-        __tablename__ = "blocked_tokens"
-
-        id = Column(Integer, primary_key=True)
-        jti = Column(String(36), nullable=False, index=True)
-
-    class SQLAlchemyDatabase(Database):
-        def __init__(self):
-            self.db_url = db_url
-            self.meta = db_meta
-            self.Base = Base
-
+    class SQLAlchemyDatabase(UserDatabase, Database):
         def init_debug(self):
-            self.meta.create_all()
-
-        @sessionmaker.with_begin
-        @from_orm(User)
-        def _create_user(self, username: str, password: str, session) -> User:
-            return UserORM.create(session, username=username, password=password)
-
-        @sessionmaker.with_begin
-        @from_orm(User)
-        def find_user(self, user_id: int, session) -> UserORM | None:
-            return session.get_first(select(UserORM).filter_by(id=user_id))
-
-        @sessionmaker.with_begin
-        @from_orm(User)
-        def find_user_by_username(self, username: str, session) -> UserORM | None:
-            return session.get_first(select(UserORM).filter_by(username=username))
-
-        @sessionmaker.with_begin
-        def block_token(self, jti: str, session) -> None:
-            return BlockedToken.create(session, jti=jti)
-
-        @sessionmaker.with_begin
-        def is_token_blocked(self, jti: str, session) -> bool:
-            return session.get_first(select(BlockedToken).filter_by(jti=jti)) is not None
+            db_meta.create_all()
 
     return SQLAlchemyDatabase()
